SDK-334 move benchmarks into own files
