+++ cpp_sdk/cpp_dist_bundle/benchmarks/CMakeLists.txt
@@ -31,7 +31,19 @@ set(TRUEFACE_SDK_DIR "${CMAKE_CURRENT_LIST_DIR}/../trueface_sdk")
 include_directories(${TRUEFACE_SDK_DIR}/include)
 link_directories(${TRUEFACE_SDK_DIR}/lib)
 
-add_executable(run_benchmarks benchmark.cpp)
+add_executable(run_benchmarks benchmark.cpp
+    benchmark_preprocess_image.cpp
+    benchmark_face_image_orientation_detection.cpp
+    benchmark_face_image_blur_detection.cpp
+    benchmark_glasses_detection.cpp
+    benchmark_face_recognition.cpp
+    benchmark_object_detection.cpp
+    benchmark_detailed_landmark_detection.cpp
+    benchmark_head_orientation.cpp
+    benchmark_mask_detection.cpp
+    benchmark_blink_detection.cpp
+    benchmark_spoof_detection.cpp
+    benchmark_face_landmark_detection.cpp)
 add_executable(run_benchmarks_1N_identification benchmark_1N_identification.cpp)
