+++ cpp_sdk/cpp_dist_bundle/benchmarks/benchmark_detailed_landmark_detection.cpp
@@ -1,0 +1,51 @@
+#include "benchmark.h"
+#include "stopwatch.h"
+
+using namespace Trueface;
+
+void benchmarkDetailedLandmarkDetection(const SDKFactory& sdkFactory, BenchmarkParams params) {
+    // Initialize the SDK
+    auto options = sdkFactory.createBasicConfiguration();
+    options.initializeModule.landmarkDetector = true;
+
+    auto tfSdk = sdkFactory.createSDK(options);
+
+    // Load the image
+    TFImage img;
+    ErrorCode errorCode = tfSdk.preprocessImage("../images/headshot.jpg", img);
+    if (errorCode != ErrorCode::NO_ERROR) {
+        std::cout << "Error: could not load the image" << std::endl;
+        return;
+    }
+
+    FaceBoxAndLandmarks faceBoxAndLandmarks;
+    bool found = false;
+    errorCode = tfSdk.detectLargestFace(img, faceBoxAndLandmarks, found);
+
+    if (errorCode != ErrorCode::NO_ERROR || !found) {
+        std::cout << "Unable to detect face in image" << std::endl;
+        return;
+    }
+
+    Landmarks landmarks;
+
+    if (params.doWarmup) {
+        for (int i = 0; i < params.numWarmup; ++i) {
+            tfSdk.getFaceLandmarks(img, faceBoxAndLandmarks, landmarks);
+            if (errorCode != ErrorCode::NO_ERROR) {
+                std::cout << "Error: Unable to get detailed landmarks" << std::endl;
+                return;
+            }
+        }
+    }
+
+    // Time the landmark detection
+    preciseStopwatch stopwatch;
+    for (size_t i = 0; i < params.numIterations; ++i) {
+        tfSdk.getFaceLandmarks(img, faceBoxAndLandmarks, landmarks);
+    }
+    auto totalTime = stopwatch.elapsedTime<float, std::chrono::milliseconds>();
+
+    std::cout << "Average time 106 face landmark detection: " << totalTime / params.numIterations
+              << " ms | " << params.numIterations << " iterations" << std::endl;
+}
