+++ cpp_sdk/cpp_dist_bundle/benchmarks/benchmark_face_image_orientation_detection.cpp
@@ -1,0 +1,42 @@
+#include "benchmark.h"
+#include "stopwatch.h"
+
+using namespace Trueface;
+
+void benchmarkFaceImageOrientationDetection(const SDKFactory& sdkFactory, BenchmarkParams params) {
+    // Initialize the SDK
+    auto options = sdkFactory.createBasicConfiguration();
+    options.initializeModule.faceOrientationDetector = true;
+
+    auto tfSdk = sdkFactory.createSDK(options);
+
+    // Load the image
+    TFImage img;
+    ErrorCode errorCode = tfSdk.preprocessImage("../images/headshot.jpg", img);
+    if (errorCode != ErrorCode::NO_ERROR) {
+        std::cout << "Error: could not load the image" << std::endl;
+        return;
+    }
+
+    RotateFlags flags;
+
+    if (params.doWarmup) {
+        for (int i = 0; i < params.numWarmup; ++i) {
+            errorCode = tfSdk.getFaceImageRotation(img, flags);
+            if (errorCode != ErrorCode::NO_ERROR) {
+                std::cout << "Error: Unable to compute face image orientation" << std::endl;
+                return;
+            }
+        }
+    }
+
+    // Time the mask detector
+    preciseStopwatch stopwatch;
+    for (size_t i = 0; i < params.numIterations; ++i) {
+        tfSdk.getFaceImageRotation(img, flags);
+    }
+    auto totalTime = stopwatch.elapsedTime<float, std::chrono::milliseconds>();
+
+    std::cout << "Average time face image orientation detection: " << totalTime / params.numIterations
+              << " ms  | " << params.numIterations << " iterations" << std::endl;
+}
