+++ cpp_sdk/cpp_dist_bundle/benchmarks/benchmark_head_orientation.cpp
@@ -1,0 +1,60 @@
+#include "benchmark.h"
+#include "stopwatch.h"
+
+using namespace Trueface;
+
+void benchmarkHeadOrientation(const SDKFactory& sdkFactory, BenchmarkParams params) {
+    // Initialize the SDK
+    auto options = sdkFactory.createBasicConfiguration();
+    options.initializeModule.faceDetector = true;
+
+    auto tfSdk = sdkFactory.createSDK(options);
+
+    // Load the image
+    TFImage img;
+    ErrorCode errorCode = tfSdk.preprocessImage("../images/headshot.jpg", img);
+    if (errorCode != ErrorCode::NO_ERROR) {
+        std::cout << "Error: could not load the image" << std::endl;
+        return;
+    }
+
+    FaceBoxAndLandmarks faceBoxAndLandmarks;
+    bool found = false;
+    errorCode = tfSdk.detectLargestFace(img, faceBoxAndLandmarks, found);
+
+    if (errorCode != ErrorCode::NO_ERROR || !found) {
+        std::cout << "Unable to detect face in image" << std::endl;
+        return;
+    }
+
+    Landmarks landmarks;
+    errorCode = tfSdk.getFaceLandmarks(img, faceBoxAndLandmarks, landmarks);
+
+    if (errorCode != ErrorCode::NO_ERROR) {
+        std::cout << "Unable to detect landmarks" << std::endl;
+        return;
+    }
+
+    float yaw, pitch, roll;
+    std::array<double, 3> rotationVec, translationVec;
+
+    if (params.doWarmup) {
+        for (int i = 0; i < params.numWarmup; ++i) {
+            errorCode = tfSdk.estimateHeadOrientation(img, faceBoxAndLandmarks, landmarks, yaw, pitch, roll, rotationVec, translationVec);
+            if (errorCode != ErrorCode::NO_ERROR) {
+                std::cout << "Unable to run head orientation method" << std::endl;
+                return;
+            }
+        }
+    }
+
+    // Time the head orientation
+    preciseStopwatch stopwatch;
+    for (size_t i = 0; i < params.numIterations; ++i) {
+        tfSdk.estimateHeadOrientation(img, faceBoxAndLandmarks, landmarks, yaw, pitch, roll, rotationVec, translationVec);
+    }
+    auto totalTime = stopwatch.elapsedTime<float, std::chrono::milliseconds>();
+
+    std::cout << "Average time head orientation: " << totalTime / params.numIterations
+              << " ms | " << params.numIterations << " iterations" << std::endl;
+}
