+++ cpp_sdk/cpp_dist_bundle/benchmarks/benchmark_mask_detection.cpp
@@ -1,0 +1,64 @@
+#include "benchmark.h"
+#include "stopwatch.h"
+
+using namespace Trueface;
+
+void benchmarkMaskDetection(const SDKFactory& sdkFactory, BenchmarkParams params) {
+    // Initialize the SDK
+    auto options = sdkFactory.createBasicConfiguration();
+    options.initializeModule.faceDetector = true;
+
+    auto tfSdk = sdkFactory.createSDK(options);
+
+    // Load the image
+    TFImage img;
+    ErrorCode errorCode = tfSdk.preprocessImage("../images/headshot.jpg", img);
+    if (errorCode != ErrorCode::NO_ERROR) {
+        std::cout << "Error: could not load the image" << std::endl;
+        return;
+    }
+
+    FaceBoxAndLandmarks faceBoxAndLandmarks;
+    bool found = false;
+    errorCode = tfSdk.detectLargestFace(img, faceBoxAndLandmarks, found);
+
+    if (errorCode != ErrorCode::NO_ERROR || !found) {
+        std::cout << "Unable to detect face in image" << std::endl;
+        return;
+    }
+
+    TFFacechip facechip;
+    errorCode = tfSdk.extractAlignedFace(img, faceBoxAndLandmarks, facechip);
+    if (errorCode != ErrorCode::NO_ERROR) {
+        std::cout << "Error: Unable to extract aligned face for mask detection" << std::endl;
+        return;
+    }
+
+    std::vector<TFFacechip> facechips;
+    for (size_t i = 0; i < params.batchSize; ++i) {
+        facechips.push_back(facechip);
+    }
+
+    std::vector<MaskLabel> maskLabels;
+    std::vector<float> maskScores;
+
+    if (params.doWarmup) {
+        for (int i = 0; i < params.numWarmup; ++i) {
+            errorCode = tfSdk.detectMasks(facechips, maskLabels, maskScores);
+            if (errorCode != ErrorCode::NO_ERROR) {
+                std::cout << "Error: Unable to run mask detection" << std::endl;
+                return;
+            }
+        }
+    }
+
+    // Time the mask detector
+    preciseStopwatch stopwatch;
+    for (size_t i = 0; i < params.numIterations; ++i) {
+        tfSdk.detectMasks(facechips, maskLabels, maskScores);
+    }
+    auto totalTime = stopwatch.elapsedTime<float, std::chrono::milliseconds>();
+
+    std::cout << "Average time mask detection: " << totalTime / params.numIterations / static_cast<float>(params.batchSize)
+              << " ms | batch size = " << params.batchSize << " | " << params.numIterations << " iterations" << std::endl;
+}
